api/requirements.txt
azure-cosmosdb-table
flask
requests
waitress